Use typedef to switch between double and single precision

+++ src/tools/args.h
@@ -105,7 +105,7 @@ public:
 		return m_size;
 	}
 
-	float timeSteps()
+	unsigned int timeSteps()
 	{
 		return m_timeSteps;
 	}
